HurtBox fixed, added up and down attack logic
Finished the bare bones player side of the melee combat system that allows the player to hit to the side or up or down (when in the air). The myenemy demonstrates how the enemies logic will react to this system.

Still need to implement the player taking damage and the enemyAI